# frontend/openapi.yaml
openapi: 3.0.3
info:
  title: Think Then Talk API
  description: The complete API specification for the Think Then Talk event platform.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /events/{eventId}:
    get:
      tags:
        - Events
      summary: Get Event Details
      description: Returns the full details for a single event.
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found.
  /admin/events:
    get:
      tags:
        - Admin
      summary: List All Events (Admin)
      description: Retrieves a list of all events for the admin panel, including past events.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of all events.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
    post:
      tags:
        - Admin
      summary: Create Event (Admin)
      description: Creates a new event.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventFormData'
      responses:
        '201':
          description: Event created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
  '/admin/events/{eventId}':
    get:
      tags:
        - Admin
      summary: Get Event (Admin)
      description: Retrieves a single event for editing in the admin panel.
      security:
        - bearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found.
    put:
      tags:
        - Admin
      summary: Update Event (Admin)
      description: Updates the core details of an event.
      security:
        - bearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventFormData'
      responses:
        '200':
          description: Event updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found.
...
  /admin/discounts:
    get:
      tags:
        - Admin
      summary: List Discounts (Admin)
      description: Retrieves a list of all discounts for the admin panel.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of all discounts.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Discount'
    post:
      tags:
        - Admin
      summary: Create Discount (Admin)
      description: Creates a new discount code or automatic offer.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscountFormData'
      responses:
        '201':
          description: Discount created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Discount'
  '/admin/discounts/{discountId}':
    put:
      tags:
        - Admin
      summary: Update Discount (Admin)
      description: Updates an existing discount.
      security:
        - bearerAuth: []
      parameters:
        - name: discountId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscountFormData'
      responses:
        '200':
          description: Discount updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Discount'
        '404':
          description: Discount not found.
    delete:
      tags:
        - Admin
      summary: Delete Discount (Admin)
      description: Deletes a discount.
      security:
        - bearerAuth: []
      parameters:
        - name: discountId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Discount deleted successfully.
        '404':
          description: Discount not found.
...
components:
  schemas:
...
    TicketType:
      type: string
      enum:
        - EARLY_BIRD
        - STANDARD
        - SUPPORTER
...
    Discount:
      type: object
      required:
        - id
        - name
        - type
        - value
        - usedCount
        - startDate
        - endDate
        - isActive
        - isPublic
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        type:
          type: string
          enum:
            - PERCENT
            - FIXED
        value:
          type: number
        applicableEventIds:
          type: array
          items:
            type: string
        maxUses:
          type: integer
        usedCount:
          type: integer
        maxUsesPerUser:
          type: integer
        minAmount:
          type: number
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        isActive:
          type: boolean
        isPublic:
          type: boolean
    DiscountFormData:
      type: object
      required:
        - name
        - type
        - value
        - startDate
        - endDate
        - isActive
        - isPublic
      properties:
        name:
          type: string
        code:
          type: string
        type:
          type: string
          enum:
            - PERCENT
            - FIXED
        value:
          type: number
        applicableEventIds:
          type: array
          items:
            type: string
        maxUses:
          type: integer
        maxUsesPerUser:
          type: integer
        minAmount:
          type: number
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        isActive:
          type: boolean
        isPublic:
          type: boolean
...
    Event:
      type: object
      required:
        - id
        - title
        - summary
        - description
        - type
        - startDateTime
        - capacityTotal
        - capacityRemaining
        - showRemainingCapacity
        - categories
        - tickets
        - resources
      properties:
        id:
          type: string
        title:
          type: string
        summary:
          type: object
          properties:
            fa:
              type: string
            en:
              type: string
        description:
          type: object
          properties:
            fa:
              type: string
            en:
              type: string
        type:
          type: string
          enum:
            - ONLINE
            - OFFLINE
        address:
          type: string
        city:
          type: string
        startDateTime:
          type: string
          format: date-time
        endDateTime:
          type: string
          format: date-time
        capacityTotal:
          type: integer
        capacityRemaining:
          type: integer
        showRemainingCapacity:
          type: boolean
        categories:
          type: array
          items:
            type: string
        tickets:
          type: array
          items:
            $ref: '#/components/schemas/EventTicketConfig'
        resources:
          type: array
          items:
            $ref: '#/components/schemas/EventResource'
        publicDiscountIds:
          type: array
          items:
            type: string
        posterUrl:
          type: string
          format: uri
    EventFormData:
      type: object
      required:
        - title
        - summary
        - description
        - type
        - startDateTime
        - capacityTotal
        - showRemainingCapacity
      properties:
        title:
          type: string
        categories:
          type: string
        summary:
          type: object
          properties:
            fa:
              type: string
            en:
              type: string
        description:
          type: object
          properties:
            fa:
              type: string
            en:
              type: string
        type:
          type: string
          enum:
            - ONLINE
            - OFFLINE
        city:
          type: string
        address:
          type: string
        startDateTime:
          type: string
          format: date-time
        capacityTotal:
          type: integer
        showRemainingCapacity:
          type: boolean
        publicDiscountIds:
          type: array
          items:
            type: string
        posterUrl:
          type: string
          format: uri
    EventResource:
      type: object
      required:
        - id
        - title
        - accessLevel
        - url
      properties:
        id:
          type: string
        title:
          type: object
          properties:
            fa:
              type: string
            en:
              type: string
        description:
          type: string
        accessLevel:
          type: string
          enum:
            - PUBLIC
            - REGISTERED_ONLY
        url:
          type: string
          format: uri
    EventTicketConfig:
      type: object
      required:
        - type
        - price
        - currency
        - quantityTotal
        - quantitySold
      properties:
        type:
          $ref: '#/components/schemas/TicketType'
        price:
          type: number
        currency:
          type: string
          enum:
            - IRR
            - TOMAN
        quantityTotal:
          type: integer
        quantitySold:
          type: integer
        earlyBirdEndDate:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT